move to start the elastic search index creation
